LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-error-1-expected.txt

CONSOLE MESSAGE: [object Object]
CONSOLE MESSAGE: [object Object]
CONSOLE MESSAGE: [object Object]
CONSOLE MESSAGE: [object Object]

PASS Test that exceptions during evaluation lead to error events on window, and that exceptions are remembered.
